requirements.txt
flask
tensorflow-cpu
pillow
numpy
gunicorn